refactor : Refactoring all codes
