+++ src/main/java/com/espinas/fhir/common/extensions/BaseDocumentModelListener.java
@@ -1,22 +1,20 @@
 package com.espinas.fhir.common.extensions;
 
-import com.espinas.fhir.domain.common.BaseEntity;
+import com.espinas.fhir.domain.common.BaseDocument;
 import com.espinas.fhir.service.common.SequenceGeneratorService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.mongodb.core.mapping.event.AbstractMongoEventListener;
 import org.springframework.data.mongodb.core.mapping.event.BeforeConvertEvent;
 import org.springframework.stereotype.Component;
 
-import java.time.LocalDateTime;
-
 @Component
 @AllArgsConstructor
-public class ValidationModelListener extends AbstractMongoEventListener<BaseEntity> {
+public class BaseDocumentModelListener extends AbstractMongoEventListener<BaseDocument> {
 
     private SequenceGeneratorService sequenceGeneratorService;
 
     @Override
-    public void onBeforeConvert(BeforeConvertEvent<BaseEntity> event) {
-        event.getSource().setId(sequenceGeneratorService.generateSequence(BaseEntity.SEQUENCE_NAME).toString());
+    public void onBeforeConvert(BeforeConvertEvent<BaseDocument> event) {
+        event.getSource().setId(sequenceGeneratorService.generateSequence(BaseDocument.SEQUENCE_NAME).toString());
     }
 }
