+++ src/main/java/com/espinas/fhir/domain/validation/collection/DatabaseSequence.java
@@ -14,5 +14,4 @@ public class DatabaseSequence {
     private String id;
 
     private Long seq;
-
 }
