+++ src/main/java/com/espinas/fhir/service/common/SequenceGeneratorService.java
@@ -19,8 +19,6 @@ public class SequenceGeneratorService {
     private final MongoOperations mongoOperations;
 
     public Long generateSequence(String seqName) {
-
-        // TODO
         DatabaseSequence counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
                 new Update().inc("seq",1), options().returnNew(true).upsert(true),
                 DatabaseSequence.class);
