--- src/test/java/com/espinas/fhir/domain/board/repository/BoardRepositoryTest.java
@@ -1,4 +1,0 @@
-//package com.espinas.fhir.domain.board.repository;
-//
-//public class BoardRepositoryTest {
-//}
